services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pyjwt==2.9.0
python-multipart==0.0.12
httpx==0.27.2
boto3==1.35.0
oss2==2.19.0
requests==2.32.3
